+++ examples/calculator-java-junit5/src/test/java/io/cucumber/examples/calculator/RunCucumberTest.java
@@ -1,13 +1,20 @@
 package io.cucumber.examples.calculator;
 
-import io.cucumber.junit.platform.engine.Cucumber;
+import org.junit.platform.suite.api.ConfigurationParameter;
+import org.junit.platform.suite.api.IncludeEngines;
+import org.junit.platform.suite.api.SelectClasspathResource;
+import org.junit.platform.suite.api.Suite;
+
+import static io.cucumber.junit.platform.engine.Constants.GLUE_PROPERTY_NAME;
 
 /**
  * Work around. Surefire does not use JUnits Test Engine discovery
- * functionality. Alternatively execute the the
+ * functionality. Alternatively execute the
  * org.junit.platform.console.ConsoleLauncher with the maven-antrun-plugin.
  */
-@Cucumber
+@Suite
+@IncludeEngines("cucumber")
+@SelectClasspathResource("io/cucumber/examples/calculator")
+@ConfigurationParameter(key = GLUE_PROPERTY_NAME, value = "io.cucumber.examples.calculator")
 public class RunCucumberTest {
-
 }