+++ junit-platform-engine/src/main/java/io/cucumber/junit/platform/engine/Cucumber.java
@@ -1,6 +1,7 @@
 package io.cucumber.junit.platform.engine;
 
 import org.apiguardian.api.API;
+import org.apiguardian.api.API.Status;
 import org.junit.platform.commons.annotation.Testable;
 
 import java.lang.annotation.ElementType;
@@ -12,21 +13,37 @@ import java.lang.annotation.Target;
  * Test discovery annotation. Marks the package of the annotated class for test
  * discovery.
  * <p>
- * Some build tools do not support the
+ * Maven and Gradle do not support the
  * {@link org.junit.platform.engine.discovery.DiscoverySelectors} used by
- * Cucumber. As a work around Cucumber will scan the package of the annotated
+ * Cucumber. As a workaround Cucumber will scan the package of the annotated
  * class for feature files and execute them.
  * <p>
  * Note about Testable: While this class is annotated with @Testable the
  * recommended way for IDEs and other tooling use the selectors implemented by
  * Cucumber to discover feature files.
+ * <p>
+ * DEPRECATED: Please use the JUnit Platform Suite to run Cucumber in
+ * combination with Surefire or Gradle. E.g: <code><pre>{@code
+ *  package com.example;
  *
- * @see CucumberTestEngine
+ *  import org.junit.platform.suite.api.ConfigurationParameter;
+ *  import org.junit.platform.suite.api.SelectClasspathResource;
+ *  import org.junit.platform.suite.api.Suite;
+ *
+ * import static io.cucumber.junit.platform.engine.Constants.GLUE_PROPERTY_NAME;
+ *  &#64;Suite
+ *  &#64;SelectClasspathResource("com/example")
+ *  &#64;ConfigurationParameter(key = GLUE_PROPERTY_NAME, value = "com.example")
+ *  public class RunCucumberTest {
+ *  }
+ * }</pre></code> @see CucumberTestEngine @deprecated Please use the JUnit
+ * Platform Suite.
  */
-@API(status = API.Status.STABLE)
+@API(status = Status.DEPRECATED)
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.TYPE)
 @Testable
+@Deprecated
 public @interface Cucumber {
 
 }