+++ kotlin-java8/src/test/kotlin/io/cucumber/kotlin/RunCucumberTest.kt
@@ -1,7 +1,13 @@
 package io.cucumber.kotlin
 
-import io.cucumber.junit.Cucumber
-import org.junit.runner.RunWith
+import io.cucumber.junit.platform.engine.Constants
+import org.junit.platform.suite.api.ConfigurationParameter
+import org.junit.platform.suite.api.IncludeEngines
+import org.junit.platform.suite.api.SelectClasspathResource
+import org.junit.platform.suite.api.Suite
 
-@RunWith(Cucumber::class)
+@Suite
+@IncludeEngines("cucumber")
+@SelectClasspathResource("io/cucumber/kotlin")
+@ConfigurationParameter(key = Constants.GLUE_PROPERTY_NAME, value = "io.cucumber.kotlin")
 class RunCucumberTest