+++ openejb/src/test/java/io/cucumber/openejb/RunCucumberTest.java
@@ -1,9 +1,16 @@
 package io.cucumber.openejb;
 
-import io.cucumber.junit.Cucumber;
-import org.junit.runner.RunWith;
+import org.junit.platform.suite.api.ConfigurationParameter;
+import org.junit.platform.suite.api.IncludeEngines;
+import org.junit.platform.suite.api.SelectClasspathResource;
+import org.junit.platform.suite.api.Suite;
 
-@RunWith(Cucumber.class)
+import static io.cucumber.junit.platform.engine.Constants.GLUE_PROPERTY_NAME;
+
+@Suite
+@IncludeEngines("cucumber")
+@SelectClasspathResource("io/cucumber/openejb")
+@ConfigurationParameter(key = GLUE_PROPERTY_NAME, value = "io.cucumber.openejb")
 public class RunCucumberTest {
 
 }