+++ spring/src/test/java/io/cucumber/spring/dirtiescontextconfig/RunCucumberTest.java
@@ -1,13 +1,16 @@
 package io.cucumber.spring.dirtiescontextconfig;
 
-import io.cucumber.junit.Cucumber;
-import io.cucumber.junit.CucumberOptions;
-import org.junit.runner.RunWith;
+import org.junit.platform.suite.api.ConfigurationParameter;
+import org.junit.platform.suite.api.IncludeEngines;
+import org.junit.platform.suite.api.SelectClasspathResource;
+import org.junit.platform.suite.api.Suite;
 
-@RunWith(Cucumber.class)
-@CucumberOptions(
-        glue = "io.cucumber.spring.dirtiescontextconfig",
-        features = "classpath:io/cucumber/spring/dirtyCukes.feature")
+import static io.cucumber.junit.platform.engine.Constants.GLUE_PROPERTY_NAME;
+
+@Suite
+@IncludeEngines("cucumber")
+@SelectClasspathResource("io/cucumber/spring/dirtyCukes.feature")
+@ConfigurationParameter(key = GLUE_PROPERTY_NAME, value = "io.cucumber.spring.dirtiescontextconfig")
 public class RunCucumberTest {
 
 }