+++ spring/src/test/java/io/cucumber/spring/metaconfig/general/RunCucumberTest.java
@@ -1,17 +1,19 @@
 package io.cucumber.spring.metaconfig.general;
 
-import io.cucumber.junit.Cucumber;
-import io.cucumber.junit.CucumberOptions;
-import org.junit.runner.RunWith;
+import org.junit.platform.suite.api.ConfigurationParameter;
+import org.junit.platform.suite.api.IncludeEngines;
+import org.junit.platform.suite.api.SelectClasspathResource;
+import org.junit.platform.suite.api.Suite;
 
-@RunWith(Cucumber.class)
-@CucumberOptions(
-        glue = {
-                "io.cucumber.spring.metaconfig.general",
-                "io.cucumber.spring.commonglue",
-                "cucumber.api.spring"
-        },
-        features = "classpath:io/cucumber/spring/springBeanInjectionWithMetaConfiguration.feature")
+import static io.cucumber.junit.platform.engine.Constants.GLUE_PROPERTY_NAME;
+
+@Suite
+@IncludeEngines("cucumber")
+@SelectClasspathResource("io/cucumber/spring/springBeanInjectionWithMetaConfiguration.feature")
+@ConfigurationParameter(
+        key = GLUE_PROPERTY_NAME,
+        value = "io.cucumber.spring.metaconfig.general," +
+                "io.cucumber.spring.commonglue")
 public class RunCucumberTest {
 
 }